SD card: add the layer between hardware and Petit FatFs.
This is all the commands to read from and write to SPI,
initializing the card, read in blocks and so on. This should
make Petit FatFs actually usable.

So far read-only and no M-codes to let end users play with
this stuff.

Code not tested on actual hardware, but it compiles and was
carefully reviewed visually, so chances are good it mostly
works.